update form userlist

+++ HotelBooking/HotelBooking/Form_UserList.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
 using System.Drawing;
+using System.Drawing.Drawing2D;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -15,11 +16,38 @@ namespace HotelBooking
     public partial class Form_UserList : Form
     {
         KhachHangBUL bul = new KhachHangBUL();
+        NhomNguoiDungBUL nhombul=new NhomNguoiDungBUL();
+        //UserButtonControl buttonControl;
+        string _username = Form_Login._username;
+        KhachHangDTO kh;
         public Form_UserList()
         {
             InitializeComponent();
             LoadUserList();
+            LoadNhomNguoiDung();
+
+            //buttonControl = new UserButtonControl();
+            //buttonControl.Location = new Point(10, 10);
+            //this.Controls.Add(buttonControl);
+
+            //buttonControl.BtnAdd_Click += btnAdd_Click;
+            //buttonControl.BtnUpdate_Click += btnUpdate_Click;
+            //buttonControl.BtnDelete_Click += btnDelete_Click;
+            //buttonControl.BtnCancel_CLick += btnCancel_Click;
         }
+    
+        
+        //private void btnDeleteImage_Click(object sender, EventArgs e)
+        //{
+        //    if (pictureBoxUser.Image != null)
+        //    {
+        //        pictureBoxUser.Image = Properties.Resources.defaultAvatar;
+        //    }
+        //    else
+        //    {
+        //        MessageBox.Show("Không có ảnh để xóa!", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+        //    }
+        //}
 
         private void LoadUserList()
         {
@@ -27,17 +55,80 @@ namespace HotelBooking
             {
                 DataTable userList = bul.GetKhachHangDataTable();
                 dgvUserList.DataSource = userList;
+                dgvUserList.ClearSelection();
+                dgvUserList.CurrentCell = null;
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show("An error occurred: " + ex.Message);
+            }
+
+        }
+        private void LoadUserList(int? userGroupId = null)
+        {
+            try
+            {
+                DataTable userList = bul.GetKhachHangDataTable(userGroupId);
+                dgvUserList.DataSource = userList;
+                dgvUserList.ClearSelection();
+                dgvUserList.CurrentCell = null;
             }
             catch (Exception ex)
             {
                 MessageBox.Show("An error occurred: " + ex.Message);
             }
         }
+
+        private void LoadNhomNguoiDung()
+        {
+            try
+            {
+                List<NhomNguoiDungDTO> nhomNguoiDungList = nhombul.GetAllNhomNguoiDung();
+                cmbUserGroup.DataSource = nhomNguoiDungList;
+                cmbUserGroup.DisplayMember = "tenNhomNguoiDung";
+                cmbUserGroup.ValueMember = "maNhomNguoiDung";
+                cmbUserGroup.SelectedIndexChanged += cmbUserGroup_SelectedIndexChanged;
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show("An error occurred: " + ex.Message);
+            }
+        }
+        //private void cmbUserGroup_SelectedIndexChanged(object sender, EventArgs e)
+        //{
+        //    if (cmbUserGroup.SelectedValue != null)
+        //    {
+        //        if (int.TryParse(cmbUserGroup.SelectedValue.ToString(), out int selectedGroupId))
+        //        {
+        //            LoadUserList(selectedGroupId);
+        //        }
+        //        else
+        //        {
+        //            LoadUserList();
+        //        }
+        //    }
+        //}
+        private void cmbUserGroup_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            if (cmbUserGroup.SelectedValue != null)
+            {
+                if (int.TryParse(cmbUserGroup.SelectedValue.ToString(), out int selectedGroupId))
+                {
+                    LoadUserList(selectedGroupId);
+                }
+                else
+                {
+                    LoadUserList();
+                }
+            }
+        }
+
         private void dgvUserList_CellClick(object sender, DataGridViewCellEventArgs e)
         {
             if (e.RowIndex >= 0)
             {
                 DataGridViewRow selected = dgvUserList.Rows[e.RowIndex];
+                
 
                 if (selected != null)
                 {
@@ -53,41 +144,103 @@ namespace HotelBooking
                     txtEmail.Text = email;
                     txtAddress.Text = diaChi;
                     txtPhone.Text = soDienThoai;
+
                 }
             }
         }
         private void btnAdd_Click(object sender, EventArgs e)
         {
-
             try
             {
-                if (string.IsNullOrWhiteSpace(txtUserName.Text) ||
-                    string.IsNullOrWhiteSpace(txtFullname.Text) ||
-                    string.IsNullOrWhiteSpace(txtEmail.Text) ||
-                    string.IsNullOrWhiteSpace(txtAddress.Text) ||
-                    string.IsNullOrWhiteSpace(txtPhone.Text))
+                if (txtUserName.Text.Trim().Length == 0)
                 {
-                    MessageBox.Show("Please fill out all fields.");
+                    MessageBox.Show("Bạn cần nhập vào username!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtUserName.Focus();
                     return;
                 }
-                KhachHangDTO newUser = new KhachHangDTO
-                {
-                    username = txtUserName.Text,
-                    hoTen = txtFullname.Text,
-                    email = txtEmail.Text,
-                    diaChi = txtAddress.Text,
-                    soDienThoai = txtPhone.Text,
-                    maNhomNguoiDung = 1 //
-                };
 
-                if (bul.AddKhachHang(newUser))
+                if (!System.Text.RegularExpressions.Regex.IsMatch(txtUserName.Text.Trim(), @"^[a-zA-Z0-9]+$"))
                 {
-                    LoadUserList();
-                    MessageBox.Show("User added successfully!");
+                    MessageBox.Show("Username không được chứa ký tự đặc biệt!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtUserName.Focus();
+                    return;
                 }
-                else
+
+                if (txtEmail.Text.Trim().Length == 0)
                 {
-                    MessageBox.Show("Failed to add user.");
+                    MessageBox.Show("Bạn cần nhập vào email!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtEmail.Focus();
+                    return;
+                }
+
+                if (!System.Text.RegularExpressions.Regex.IsMatch(txtEmail.Text.Trim(), @"^[^@\s]+@[^@\s]+\.[^@\s]+$"))
+                {
+                    MessageBox.Show("Email không hợp lệ!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtEmail.Focus();
+                    return;
+                }
+
+                if (txtFullname.Text.Trim().Length == 0)
+                {
+                    MessageBox.Show("Bạn cần nhập vào fullname!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtFullname.Focus();
+                    return;
+                }
+
+                if (txtAddress.Text.Trim().Length == 0)
+                {
+                    MessageBox.Show("Bạn cần nhập vào address!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtAddress.Focus();
+                    return;
+                }
+
+                if (txtPhone.Text.Trim().Length == 0)
+                {
+                    MessageBox.Show("Bạn cần nhập vào phone!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtPhone.Focus();
+                    return;
+                }
+
+                if (!System.Text.RegularExpressions.Regex.IsMatch(txtPhone.Text.Trim(), @"^\d+$") || txtPhone.Text.Trim().Length != 10 || !txtPhone.Text.Trim().StartsWith("0"))
+                {
+                    MessageBox.Show("Số điện thoại không hợp lệ!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtPhone.Focus();
+                    return;
+                }
+
+                string username = txtUserName.Text.Trim();
+                string password = "12345678";
+                string hoten = txtFullname.Text.Trim();
+                string diachi = txtAddress.Text.Trim();
+                string email = txtEmail.Text.Trim();
+                string sdt = txtPhone.Text.Trim();
+
+                int maNhomNguoiDung = (int)cmbUserGroup.SelectedValue;
+                bool _isSuccess = bul.RegisterNewUser(username, password, hoten, email, diachi, sdt);
+
+                if (_isSuccess)
+                {
+                    KhachHangDTO newUser = new KhachHangDTO
+                    {
+                        username = username,
+                        hoTen = hoten,
+                        password = password,
+                        email = email,
+                        diaChi = diachi,
+                        soDienThoai = sdt,
+                        maNhomNguoiDung = maNhomNguoiDung
+                    };
+
+                    bool isSuccess = bul.AddKhachHang(newUser);
+                    if (isSuccess)
+                    {
+                        LoadUserList(maNhomNguoiDung);
+                        MessageBox.Show("User added successfully!");
+                    }
+                    else
+                    {
+                        MessageBox.Show("Failed to add user.");
+                    }
                 }
             }
             catch (Exception ex)
@@ -95,13 +248,71 @@ namespace HotelBooking
                 MessageBox.Show("An error occurred: " + ex.Message);
             }
         }
+
         private void btnUpdate_Click(object sender, EventArgs e)
         {
             try
             {
+                if (txtUserName.Text.Trim().Length == 0)
+                {
+                    MessageBox.Show("Bạn cần nhập vào username!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtUserName.Focus();
+                    return;
+                }
+
+                if (!System.Text.RegularExpressions.Regex.IsMatch(txtUserName.Text.Trim(), @"^[a-zA-Z0-9]+$"))
+                {
+                    MessageBox.Show("Username không được chứa ký tự đặc biệt!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtUserName.Focus();
+                    return;
+                }
+
+                if (txtEmail.Text.Trim().Length == 0)
+                {
+                    MessageBox.Show("Bạn cần nhập vào email!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtEmail.Focus();
+                    return;
+                }
+
+                if (!System.Text.RegularExpressions.Regex.IsMatch(txtEmail.Text.Trim(), @"^[^@\s]+@[^@\s]+\.[^@\s]+$"))
+                {
+                    MessageBox.Show("Email không hợp lệ!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtEmail.Focus();
+                    return;
+                }
+
+                if (txtFullname.Text.Trim().Length == 0)
+                {
+                    MessageBox.Show("Bạn cần nhập vào fullname!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtFullname.Focus();
+                    return;
+                }
+
+                if (txtAddress.Text.Trim().Length == 0)
+                {
+                    MessageBox.Show("Bạn cần nhập vào address!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtAddress.Focus();
+                    return;
+                }
+
+                if (txtPhone.Text.Trim().Length == 0)
+                {
+                    MessageBox.Show("Bạn cần nhập vào phone!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtPhone.Focus();
+                    return;
+                }
+
+                if (!System.Text.RegularExpressions.Regex.IsMatch(txtPhone.Text.Trim(), @"^\d+$") || txtPhone.Text.Trim().Length != 10 || !txtPhone.Text.Trim().StartsWith("0"))
+                {
+                    MessageBox.Show("Số điện thoại không hợp lệ!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    txtPhone.Focus();
+                    return;
+                }
                 if (dgvUserList.SelectedRows.Count > 0)
                 {
                     int maKH = (int)dgvUserList.SelectedRows[0].Cells["ID"].Value;
+
+                    int maNhomNguoiDung = (int)cmbUserGroup.SelectedValue;
 
                     KhachHangDTO updatedUser = new KhachHangDTO
                     {
@@ -111,11 +322,11 @@ namespace HotelBooking
                         email = txtEmail.Text,
                         diaChi = txtAddress.Text,
                         soDienThoai = txtPhone.Text,
-                        maNhomNguoiDung = 1 // 
+                        maNhomNguoiDung = maNhomNguoiDung 
                     };
 
                     bul.UpdateKhachHang(updatedUser);
-                    LoadUserList();
+                    LoadUserList(maNhomNguoiDung);
                     MessageBox.Show("User updated successfully!");
                 }
                 else
@@ -133,17 +344,27 @@ namespace HotelBooking
         {
             try
             {
-                if (dgvUserList.SelectedRows.Count > 0)
+                DialogResult result = MessageBox.Show(
+                    "Bạn có chắc chắn muốn xóa không?",
+                    "Xác nhận",
+                    MessageBoxButtons.YesNo,
+                    MessageBoxIcon.Question
+                );
+                if (result == DialogResult.Yes)
                 {
-                    int maKH = (int)dgvUserList.SelectedRows[0].Cells["ID"].Value;
+                    if (dgvUserList.SelectedRows.Count > 0)
+                    {
+                        int maKH = (int)dgvUserList.SelectedRows[0].Cells["ID"].Value;
 
-                    bul.DeleteKhachHang(maKH);
-                    LoadUserList();
-                    MessageBox.Show("User deleted successfully!");
-                }
-                else
-                {
-                    MessageBox.Show("Please select a user to delete.");
+                        bul.DeleteKhachHang(maKH);
+                        int maNhomNguoiDung = (int)cmbUserGroup.SelectedValue;
+                        LoadUserList(maNhomNguoiDung); 
+                        MessageBox.Show("User deleted successfully!");
+                    }
+                    else
+                    {
+                        MessageBox.Show("Please select a user to delete.");
+                    }
                 }
             }
             catch (Exception ex)
@@ -151,23 +372,22 @@ namespace HotelBooking
                 MessageBox.Show("An error occurred: " + ex.Message);
             }
         }
-
-        private void btnReturn_Click(object sender, EventArgs e)
-        {
-            txtUserName.Clear();
-            txtFullname.Clear();
-            txtEmail.Clear();
-            txtAddress.Clear();
-            txtPhone.Clear();
-        }
-
         private void btnCancel_Click(object sender, EventArgs e)
         {
-            txtUserName.Clear();
-            txtFullname.Clear();
-            txtEmail.Clear();
-            txtPhone.Clear();
-            txtAddress.Clear();
+            DialogResult result = MessageBox.Show(
+                         "Bạn có chắc chắn muốn hủy không?",
+                         "Xác nhận",
+                         MessageBoxButtons.YesNo,
+                         MessageBoxIcon.Question
+                     );
+            if (result == DialogResult.Yes)
+            {
+                txtUserName.Clear();
+                txtFullname.Clear();
+                txtEmail.Clear();
+                txtPhone.Clear();
+                txtAddress.Clear();
+            }
         }
     }
 }
